webui: tests: extend the TestInstallationProgress test to include reboot
For making this possible add --wait option to virt-install invocation.
Read bellow for explanation:

When doing an install, one can't have the guest reboot at completion.
The initial guest configuration for launching the installer is different,
from the configuration for launching it post-boot.
Thus virt-install "fakes" the reboot by waiting for the guest to exit,
and then manually restarts it.  When using --noautoconsole this won't
happen. One needs to use --wait to workaround that. (explanation taken from [0])

[0] https://listman.redhat.com/archives/virt-tools-list/2011-November/002869.html

Co-authored-by: Katerina Koukiou <kkoukiou@redhat.com>