workflows: Trigger anaconda Web UI integration tests on pyanaconda/** changes
We want to make sure to not break the Web UI with changes which affect it,
i.e. API changes.

Add a workflow which runs if the PR affects Anaconda (changes to pyanaconda),
polls the packit COPR until it has the current PR version available, and
then test-triggers a "anaconda PR" scenario.